Fix split package in analysis-icu plugin
The analysis-icu plugin reuses server package names for analysis and
mappers. This commit moves the plugin implementation to use a single
package name, o.e.p.analysis.icu
