+++ plugins/analysis-icu/src/main/java/org/elasticsearch/plugin/analysis/icu/AnalysisICUPlugin.java
@@ -12,16 +12,8 @@ import org.apache.lucene.analysis.Analyzer;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.index.analysis.AnalyzerProvider;
 import org.elasticsearch.index.analysis.CharFilterFactory;
-import org.elasticsearch.index.analysis.IcuAnalyzerProvider;
-import org.elasticsearch.index.analysis.IcuCollationTokenFilterFactory;
-import org.elasticsearch.index.analysis.IcuFoldingTokenFilterFactory;
-import org.elasticsearch.index.analysis.IcuNormalizerCharFilterFactory;
-import org.elasticsearch.index.analysis.IcuNormalizerTokenFilterFactory;
-import org.elasticsearch.index.analysis.IcuTokenizerFactory;
-import org.elasticsearch.index.analysis.IcuTransformTokenFilterFactory;
 import org.elasticsearch.index.analysis.TokenFilterFactory;
 import org.elasticsearch.index.analysis.TokenizerFactory;
-import org.elasticsearch.index.mapper.ICUCollationKeywordFieldMapper;
 import org.elasticsearch.index.mapper.Mapper;
 import org.elasticsearch.indices.analysis.AnalysisModule.AnalysisProvider;
 import org.elasticsearch.plugins.AnalysisPlugin;
