+++ plugins/analysis-icu/src/main/java/org/elasticsearch/plugin/analysis/icu/IndexableBinaryStringTools.java
@@ -1,4 +1,12 @@
-package org.elasticsearch.index.analysis;
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+package org.elasticsearch.plugin.analysis.icu;
 
 /* @notice
  * Licensed to the Apache Software Foundation (ASF) under one or more
