+++ plugins/analysis-icu/src/test/java/org/elasticsearch/plugin/analysis/icu/IcuTokenizerFactoryTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.plugin.analysis.icu;
 
 import org.apache.lucene.analysis.icu.segmentation.ICUTokenizer;
 import org.elasticsearch.Version;
@@ -14,7 +14,7 @@ import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.Index;
-import org.elasticsearch.plugin.analysis.icu.AnalysisICUPlugin;
+import org.elasticsearch.index.analysis.TokenizerFactory;
 import org.elasticsearch.test.ESTestCase;
 
 import java.io.IOException;
@@ -81,7 +81,7 @@ public class IcuTokenizerFactoryTests extends ESTestCase {
         Files.copy(keywords, config.resolve("KeywordTokenizer.rbbi"));
         Files.copy(latin, config.resolve("Latin-dont-break-on-hyphens.rbbi"));
 
-        String json = "/org/elasticsearch/index/analysis/icu_analysis.json";
+        String json = "/org/elasticsearch/plugin/analysis/icu/icu_analysis.json";
 
         Settings settings = Settings.builder()
             .loadFromStream(json, IcuTokenizerFactoryTests.class.getResourceAsStream(json), false)
