+++ plugins/analysis-icu/src/test/resources/org/elasticsearch/plugin/analysis/icu/KeywordTokenizer.rbbi
