+++ plugins/analysis-icu/src/test/resources/org/elasticsearch/plugin/analysis/icu/Latin-dont-break-on-hyphens.rbbi
